# service/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  metadataController.AddDevelopRequest:
    properties:
      category:
        example: กล้ามเนื้อมัดใหญ่
        type: string
      detail:
        example: การนั่ง
        type: string
      image:
        example: https://res.cloudinary.com/demo/image/upload/sample.jpg
        type: string
      note:
        example: เด็กจะเริ่มนั่งได้เองเมื่ออายุประมาณ 6 เดือน โดยไม่ต้องพิง
        type: string
    required:
    - category
    - detail
    - image
    - note
    type: object
...
  metadataController.GetByCategoryAndDetailRequest:
    properties:
      category:
        example: กล้ามเนื้อมัดใหญ่
        type: string
      detail:
        example: การนั่ง
        type: string
    required:
    - category
    - detail
    type: object
  metadataController.GetByCategoryRequest:
    properties:
      category:
        example: กล้ามเนื้อมัดใหญ่
        type: string
    required:
    - category
    type: object
  metadataController.UpdateDevelopRequest:
    properties:
      category:
        example: การเคลื่อนไหว
        type: string
      detail:
        example: การนั่ง
        type: string
      image:
        example: https://res.cloudinary.com/demo/image/upload/sample.jpg
        type: string
      note:
        example: เด็กจะเริ่มนั่งได้เองเมื่ออายุประมาณ 6 เดือน โดยไม่ต้องพิง
        type: string
    type: object
...
  metadataModel.MetaDevelop:
    properties:
      category:
        type: string
      detail:
        type: string
      id:
        type: string
      image:
        type: string
      note:
        type: string
    type: object
...
host: localhost:5000
info:
  contact: {}
  description: Backend API for MimiCare Project (Gin + MongoDB)
  title: MimiCare API
  version: "1.0"
paths:
...
  /metaDevelop/add:
    post:
      consumes:
      - application/json
      description: ใช้สำหรับเพิ่มข้อมูลพัฒนาการ (admin เท่านั้น)
      parameters:
      - description: ข้อมูลช่วงอายุ
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/metadataController.AddDevelopRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties: true
            type: object
      summary: เพิ่มพัฒนาการใหม่
      tags:
      - MetaDevelop
  /metaDevelop/delete/{id}:
    delete:
      description: ลบข้อมูลพัฒนาการตาม ObjectID (admin เท่านั้น)
      parameters:
      - description: ObjectID ของพัฒนาการ
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: ลบข้อมูลพัฒนาการตาม ID
      tags:
      - MetaDevelop
  /metaDevelop/get-all:
    get:
      description: คืนค่าข้อมูลพัฒนาการทั้งหมด (admin เท่านั้น)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/metadataModel.MetaDevelop'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      summary: ดึงข้อมูลพัฒนาการทั้งหมด
      tags:
      - MetaDevelop
  /metaDevelop/get-by-category:
    post:
      consumes:
      - application/json
      description: คืนค่าข้อมูลพัฒนาการตามหมวดหมู่ (admin เท่านั้น)
      parameters:
      - description: หมวดหมู่พัฒนาการ
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/metadataController.GetByCategoryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/metadataModel.MetaDevelop'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      summary: ดึงข้อมูลพัฒนาการตามหมวดหมู่
      tags:
      - MetaDevelop
  /metaDevelop/get-by-category-detail:
    post:
      consumes:
      - application/json
      description: คืนค่าข้อมูลพัฒนาการตามหมวดหมู่และรายละเอียด (admin เท่านั้น)
      parameters:
      - description: หมวดหมู่และรายละเอียดพัฒนาการ
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/metadataController.GetByCategoryAndDetailRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/metadataModel.MetaDevelop'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      summary: ดึงข้อมูลพัฒนาการตามหมวดหมู่และรายละเอียด
      tags:
      - MetaDevelop
  /metaDevelop/get/{id}:
    get:
      description: คืนค่าข้อมูลพัฒนาการตาม ObjectID (admin เท่านั้น)
      parameters:
      - description: ObjectID ของพัฒนาการ
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/metadataModel.MetaDevelop'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      summary: ดึงข้อมูลพัฒนาการตาม ID
      tags:
      - MetaDevelop
  /metaDevelop/update/{id}:
    put:
      consumes:
      - application/json
      description: ใช้สำหรับอัปเดตข้อมูลพัฒนาการตาม ObjectID (admin เท่านั้น)
      parameters:
      - description: ObjectID ของพัฒนาการที่ต้องการอัปเดต
        in: path
        name: id
        required: true
        type: string
      - description: ข้อมูลพัฒนาการที่ต้องการอัปเดต
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/metadataController.UpdateDevelopRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/metadataModel.MetaDevelop'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: อัปเดตข้อมูลพัฒนาการตาม ID
      tags:
      - MetaDevelop
...
schemes:
- http
swagger: "2.0"